+++ CHANGELOG.md
@@ -3,5 +3,23 @@ All notable changes to the "leet-html" extension will be documented in this file
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
-- Initial release
+## Changes
+- 1.1.1 Initial release
+- 1.1.2 Add support for 
+    ```js
+    htm()`<div class="abc">Some Content</div>`
+    html()`<div class="abc">Some Content</div>`
+    htmx()`<div class="abc">Some Content</div>`
+    svg()`<div class="abc">Some Content</div>`
+    s()`<div class="abc">Some Content</div>`
+    h()`<div class="abc">Some Content</div>`
+    ``` 
+    and
+    ```js
+    /*htm*/`<div class="abc">Some Content</div>`
+    /*html*/`<div class="abc">Some Content</div>`
+    /*htmx*/`<div class="abc">Some Content</div>`
+    /*svg*/`<div class="abc">Some Content</div>`
+    /*s*/`<div class="abc">Some Content</div>`
+    /*h*/`<div class="abc">Some Content</div>`
+    ```